requirements.txt
raganything
lightrag-hku
openai
numpy
tqdm
python-dotenv
sentence-transformers
pydub
sounddevice
scipy
faster-whisper
pyttsx3
prometheus-client


opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
python-json-logger
psutil